+++ lib/routes/app_pages.dart
@@ -8,7 +8,7 @@ import 'package:solution_challenge/view/login/register/register_complete_page.da
 import 'package:solution_challenge/view/login/register/register_name_page.dart';
 import 'package:solution_challenge/view/login/register/register_number_page.dart';
 import 'package:solution_challenge/view/login/register/register_type_page.dart';
-import 'package:solution_challenge/view/my_page.dart';
+import 'package:solution_challenge/view/mypage/my_page.dart';
 part './app_routes.dart';
 
 // 앱 내 페이지를 정의하는 공간, 페이지 만든 후 여기에 정의해주세요
