+++ lib/view/home/home_page.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:solution_challenge/view/home/home_subpages/today_todo_page.dart';
-import 'package:solution_challenge/view/my_page.dart';
-import 'package:solution_challenge/view/test.dart';
+import 'package:solution_challenge/view/mypage/my_page.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({super.key});
