+++ lib/view/mypage/manage_my_account/my_account_page.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+//import 'package:get/get.dart';
+
+class MyAccountManage extends StatelessWidget {
+  const MyAccountManage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Text('hello');
+  }
+}
